fix util types
