+++ src/utils/document.service.ts
@@ -1,6 +1,7 @@
 import { Injectable, OnDestroy } from "@angular/core";
 import { Observable, Subscription } from "rxjs";
-import { EventHandler, getEventObservable } from "./event.service";
+import { EventHandler } from "./types";
+import { getEventObservable } from "./event-observable";
 
 @Injectable({
 	providedIn: "root"
