+++ src/utils/types.ts
@@ -1,0 +1,1 @@
+export type EventHandler = (event: Event) => void;
